manual backport of #2842

+++ raster/r.cost/main.c
@@ -517,7 +517,7 @@ int main(int argc, char *argv[])
 	costs.cost_out = dnullval;
 	costs.nearest = 0;
 
-	total_cells = nrows * ncols;
+	total_cells = (long)nrows * ncols;
 
 	skip_nulls = Rast_is_d_null_value(&null_cost);
 
@@ -809,7 +809,7 @@ int main(int argc, char *argv[])
      */
 
     G_debug(1, "total cells: %ld", total_cells);
-    G_debug(1, "nrows x ncols: %d", nrows * ncols);
+    G_debug(1, "nrows x ncols: %ld", (long)nrows * ncols);
     G_message(_("Finding cost path..."));
     n_processed = 0;
     visited = flag_create(nrows, ncols);
